packages/backend/pyproject.toml
[project]
name = "obsidian-search"
version = "0.1.0"
description = "Semantic search backend for Obsidian vaults"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=3.0",
    "sqlite-vec>=0.1",
    "pymupdf4llm>=0.0.17",
    "trafilatura>=1.12",
    "markdown-it-py>=3.0",
    "python-frontmatter>=1.1",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "fastmcp>=2.0",
    "watchdog>=4.0",
    "pydantic-settings>=2.0",
    "nltk>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "mypy>=1.11",
    "ruff>=0.6",
    "pre-commit>=3.8",
]

[project.scripts]
obsidian-search-api = "obsidian_search.api.server:main"
obsidian-search-mcp = "obsidian_search.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_search"]
